% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coerce-RegionSetDE.counts-DGEList}
\alias{coerce-RegionSetDE.counts-DGEList}
\alias{coerce,RegionSetDE.counts,DGEList-method}
\title{Coerce a counts object to a DGEList}
\value{
A \code{DGEList} holding the counts, the sample metadata, the region annotation and the normalisation as offsets.
}
\description{
Registers the \code{as(x, "DGEList")} idiom, which calls \code{\link{asDGEList}} with its defaults.
}
\examples{
counts <- loadExampleData("counts", verbose = FALSE)
counts <- normalizeCounts(counts, method = "background", verbose = FALSE)

dgeList <- as(counts, "DGEList")
dgeList$samples

}
\seealso{
\code{\link{asDGEList}}
}
\author{
Sebastian Gregoricchio
}
