% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConsensusPeaks.R
\name{consensusData}
\alias{consensusData}
\alias{consensusData,RegionSetDE-method}
\title{consensusData}
\usage{
consensusData(object)

\S4method{consensusData}{RegionSetDE}(object)
}
\arguments{
\item{object}{\code{RegionSetDE} object.}
}
\value{
A list with \code{groups}, the consensus regions of every group; \code{objects}, the \code{consensusRegions} object behind each of them (\code{NULL} for the groups with a single sample); \code{total}, the pooled consensus; \code{peaks}, the peaks of every sample after the blacklist and the greylist; \code{removed}, the peaks they took out, with the \code{sample} they came from and the list that removed them in \code{removed.by}; \code{samples}, a data.frame with the group, the peak file, the number of peaks read (\code{n.peaks}), removed by the blacklist (\code{n.blacklist}), by the greylist (\code{n.greylist}) and by both (\code{n.excluded}) for every sample; \code{sheet}, the sample sheet the consensus was built from; \code{groupBy}; \code{mode}, one among \code{"consensus"}, \code{"split"} and \code{"replace"}; and \code{blacklist} and \code{greylist}, the regions the peaks were cleaned against, \code{NULL} when a list was not given.
}
\description{
Returns the consensus data kept by a \code{RegionSetDE} object built from peaks with \code{\link{loadConsensusPeaks}}.
}
\examples{
if (requireNamespace("consensusRegions", quietly = TRUE)) {
  peakFiles <- list.files(system.file("extdata", package = "consensusRegions"),
                          pattern = "rep[0-9]\\\\.narrowPeak$", full.names = TRUE)

  sampleSheet <- loadSampleSheet(data.frame(sample = c("A_1", "A_2", "B_1", "B_2"),
                                            bam = c("A_1.bam", "A_2.bam", "B_1.bam", "B_2.bam"),
                                            peaks = peakFiles[c(1, 2, 2, 3)],
                                            condition = c("A", "A", "B", "B")),
                                 checkFiles = FALSE, verbose = FALSE)

  regions <- loadConsensusPeaks(sampleSheet, groupBy = "condition", verbose = FALSE)

  consensusData(regions)$samples
  lengths(consensusData(regions)$groups)
}

}
\seealso{
\code{\link{loadConsensusPeaks}}, \code{\link{consensusGroupList}}, \code{\link{plotPeakUpset}}
}
\author{
Sebastian Gregoricchio
}
