% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMethods.R
\name{contrastName}
\alias{contrastName}
\alias{contrastName,RegionSetDE.results-method}
\title{contrastName}
\usage{
contrastName(results)

\S4method{contrastName}{RegionSetDE.results}(results)
}
\arguments{
\item{results}{\code{RegionSetDE.results} object.}
}
\value{
A string.
}
\description{
Returns the contrast that produced a \code{RegionSetDE.results} object.
}
\author{
Sebastian Gregoricchio
}
