% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{countingLevel}
\alias{countingLevel}
\alias{countingLevel,RegionSetDE.counts-method}
\alias{countingLevel,RegionSetDE.fit-method}
\alias{countingLevel,RegionSetDE.results-method}
\title{countingLevel}
\usage{
countingLevel(object)

\S4method{countingLevel}{RegionSetDE.counts}(object)

\S4method{countingLevel}{RegionSetDE.fit}(object)

\S4method{countingLevel}{RegionSetDE.results}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.counts}, \code{RegionSetDE.fit} or \code{RegionSetDE.results} object.}
}
\value{
Either \code{"region"} or \code{"tile"}.
}
\description{
Tells whether the rows of an object are whole regions or tiles of a region.
}
\examples{
counts <- loadExampleData("counts", verbose = FALSE)
countingLevel(counts)

fit <- loadExampleData("fit", verbose = FALSE)
countingLevel(fit)

}
\seealso{
\code{\link{countReads}}, \code{\link{tileTable}}
}
\author{
Sebastian Gregoricchio
}
