% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRegions.R
\name{.asFormula}
\alias{.asFormula}
\title{.asFormula}
\usage{
.asFormula(x, parameterName = "design")
}
\arguments{
\item{x}{Object passed by the user.}

\item{parameterName}{String with the name of the parameter, used in the error message.}
}
\value{
A formula when the input was a string, the input itself otherwise.
}
\description{
Parses a design written as a string, leaving anything else untouched.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
