% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamUtils.R
\name{.bamIndexPath}
\alias{.bamIndexPath}
\title{.bamIndexPath}
\usage{
.bamIndexPath(bamFile)
}
\arguments{
\item{bamFile}{String with the path of the BAM file.}
}
\value{
String with the path of the first index found, \code{NA} when the file has none.
}
\description{
Looks for the index of a BAM file under the four names one can go under: \code{file.bam.bai}, \code{file.bai}, \code{file.bam.csi} and \code{file.csi}.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
