% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRegionSets.R
\name{.checkWithinSetPolicy}
\alias{.checkWithinSetPolicy}
\title{.checkWithinSetPolicy}
\usage{
.checkWithinSetPolicy(overlapWithinSet)
}
\arguments{
\item{overlapWithinSet}{Value passed by the user.}
}
\value{
The value in lower case.
}
\description{
Checks the value of \code{overlapWithinSet}.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
