% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRegions.R
\name{.fitDESeq2}
\alias{.fitDESeq2}
\title{.fitDESeq2}
\usage{
.fitDESeq2(
  countMatrix,
  designMatrix,
  offsetMatrix,
  librarySizes,
  colTable,
  verbose = TRUE
)
}
\arguments{
\item{countMatrix}{Integer matrix of counts.}

\item{designMatrix}{Design matrix.}

\item{offsetMatrix}{Matrix of log offsets, or \code{NULL}.}

\item{librarySizes}{Numeric vector with the library sizes.}

\item{colTable}{Data.frame with the sample metadata.}

\item{verbose}{Logical value to indicate whether the messages must be printed.}
}
\value{
A list with the \code{fit}, the \code{blocking} and the \code{dispersion} elements.
}
\description{
Fits the negative binomial model of \code{DESeq2} on a user supplied model matrix.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
