% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRegions.R
\name{.fitOffsets}
\alias{.fitOffsets}
\title{.fitOffsets}
\usage{
.fitOffsets(counts, useOffsets = TRUE, verbose = TRUE)
}
\arguments{
\item{counts}{\code{RegionSetDE.counts} object.}

\item{useOffsets}{Logical value indicating whether the normalisation must be used.}

\item{verbose}{Logical value to indicate whether the messages must be printed.}
}
\value{
A numeric matrix of log offsets, or \code{NULL}.
}
\description{
Extracts the normalisation stored in a counts object as a matrix of log offsets.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
