% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRegionSets.R
\name{.interRegionCor}
\alias{.interRegionCor}
\title{.interRegionCor}
\usage{
.interRegionCor(expressionMatrix, design, index)
}
\arguments{
\item{expressionMatrix}{Numeric matrix of log2 values.}

\item{design}{Design matrix.}

\item{index}{Integer vector with the rows of the set.}
}
\value{
A numeric value.
}
\description{
Estimates the correlation between the regions of a set, from the residuals of the design.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
