% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSets.R
\name{.lightenColour}
\alias{.lightenColour}
\title{.lightenColour}
\usage{
.lightenColour(colour, amount = 0.5)
}
\arguments{
\item{colour}{String with the colour.}

\item{amount}{Numeric value between 0 and 1 with the fraction of white.}
}
\value{
A string with the mixed colour.
}
\description{
Mixes a colour with white, used to build the light end of the shades of a group.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
