% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{.orderSampleValues}
\alias{.orderSampleValues}
\title{.orderSampleValues}
\usage{
.orderSampleValues(values, sampleNames, parameterName)
}
\arguments{
\item{values}{Numeric vector, named or not.}

\item{sampleNames}{Character vector with the sample names, in the order of the columns.}

\item{parameterName}{String with the name of the parameter, used in the error messages.}
}
\value{
The numeric vector, ordered as the columns of the object.
}
\description{
Checks a vector of user supplied values against the samples of a counts object and puts it in the order of the columns.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
