% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{.profileScaling}
\alias{.profileScaling}
\title{.profileScaling}
\usage{
.profileScaling(counts, useOffsets, verbose)
}
\arguments{
\item{counts}{\code{RegionSetDE.counts} object.}

\item{useOffsets}{Logical value indicating whether the scaling factors of the normalisation must be used.}

\item{verbose}{Logical value to indicate whether the messages must be printed.}
}
\value{
A numeric vector with one divisor per sample.
}
\description{
Returns the divisor turning the coverage of every sample into coverage per million fragments of the average library, scaled by the normalisation of the object when there is one.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
