% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFragmentLength.R
\name{.regionsAsGRanges}
\alias{.regionsAsGRanges}
\title{.regionsAsGRanges}
\usage{
.regionsAsGRanges(regions)
}
\arguments{
\item{regions}{A \code{RegionSetDE} object, a \code{GRangesList}, a list of \code{GRanges} or a \code{GRanges}.}
}
\value{
A \code{GRanges}.
}
\description{
Pools the regions of a \code{RegionSetDE} object, a \code{GRangesList} or a list of \code{GRanges} into a single \code{GRanges}, without metadata.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
