% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionLists.R
\name{.resolveGenomeName}
\alias{.resolveGenomeName}
\title{.resolveGenomeName}
\usage{
.resolveGenomeName(genome)
}
\arguments{
\item{genome}{String with the assembly given by the user.}
}
\value{
A string with the name used in the index, the input itself when it is not a known alias.
}
\description{
Turns the name of an assembly into the one the files are indexed under, so that GRCh38 and hg38 reach the same list.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
