% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConsensusPeaks.R
\name{.runGroupConsensus}
\alias{.runGroupConsensus}
\title{.runGroupConsensus}
\usage{
.runGroupConsensus(groupArguments)
}
\arguments{
\item{groupArguments}{List with the arguments of \code{consensusRegions::runConsensus} for the group, the peaks included.}
}
\value{
A \code{consensusRegions} object.
}
\description{
Builds the consensus of one group through \code{consensusRegions::runConsensus}, called by the workers of \code{\link{loadConsensusPeaks}}.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
