% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastInfo.R
\name{.singleContrastInfo}
\alias{.singleContrastInfo}
\title{.singleContrastInfo}
\usage{
.singleContrastInfo(results, contrastKey)
}
\arguments{
\item{results}{\code{RegionSetDE.results} object.}

\item{contrastKey}{String with the name the contrast goes under.}
}
\value{
A data.frame with one row.
}
\description{
Builds the row of \code{\link{contrastInfo}} for one contrast.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
