% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{.stackProfileRegions}
\alias{.stackProfileRegions}
\title{.stackProfileRegions}
\usage{
.stackProfileRegions(groupList)
}
\arguments{
\item{groupList}{List of \code{GRanges}, one per group of rows.}
}
\value{
A \code{GRanges}.
}
\description{
Stacks the groups of rows of a profile into one \code{GRanges}, dropping their seqinfo so that groups from different sources can be pasted together.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
