% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRegions.R
\name{.widthStratum}
\alias{.widthStratum}
\title{.widthStratum}
\usage{
.widthStratum(rowWidths, strataNumber = 5)
}
\arguments{
\item{rowWidths}{Numeric vector with the widths.}

\item{strataNumber}{Numeric value with the number of strata.}
}
\value{
A character vector with the stratum of every row.
}
\description{
Splits a vector of widths into strata of comparable size, used to keep the relative filters from favouring the wide regions.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
