% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{.writeTable}
\alias{.writeTable}
\title{.writeTable}
\usage{
.writeTable(table, path, compress = TRUE)
}
\arguments{
\item{table}{Data.frame to write.}

\item{path}{String with the destination.}

\item{compress}{Logical value indicating whether the file must be gzipped.}
}
\value{
Invisibly the path.
}
\description{
Writes a table as a tab separated file, gzipped when asked for.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
