% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{filteringLog}
\alias{filteringLog}
\alias{filteringLog,RegionSetDE.provenance-method}
\title{filteringLog}
\usage{
filteringLog(object)

\S4method{filteringLog}{RegionSetDE.provenance}(object)
}
\arguments{
\item{object}{Any object of the package: \code{RegionSetDE}, \code{RegionSetDE.counts}, \code{RegionSetDE.fit}, \code{RegionSetDE.results}, \code{RegionSetDE.setResults} or \code{RegionSetDE.setScores}.}
}
\value{
A data.frame with one row per filtering step, empty when no filter has been applied.
}
\description{
Returns the record of the filters applied to a RegionSetDE object, with the number of regions before and after each step. The record travels with the regions into the counts, the fit and the results.
}
\examples{
regionTable <- loadExampleData("regions", verbose = FALSE)
exclusionRegions <- loadExampleData("exclusionRegions", verbose = FALSE)

regions <- splitLoadRegions(GenomicRanges::makeGRangesFromDataFrame(regionTable, keep.extra.columns = TRUE),
                            splitBy = "setName", genomeAssembly = "rn4", verbose = FALSE)
regions <- applyBlacklist(regions, blacklist = exclusionRegions, verbose = FALSE)

filteringLog(regions)

}
\seealso{
\code{\link{applyBlacklist}}, \code{\link{applyWhitelist}}, \code{\link{applyGreylist}}
}
\author{
Sebastian Gregoricchio
}
