% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{genomeAssembly}
\alias{genomeAssembly}
\alias{genomeAssembly,RegionSetDE.provenance-method}
\title{genomeAssembly}
\usage{
genomeAssembly(object)

\S4method{genomeAssembly}{RegionSetDE.provenance}(object)
}
\arguments{
\item{object}{Any object of the package: \code{RegionSetDE}, \code{RegionSetDE.counts}, \code{RegionSetDE.fit}, \code{RegionSetDE.results}, \code{RegionSetDE.setResults} or \code{RegionSetDE.setScores}.}
}
\value{
A string with the assembly, or \code{NULL} when none was declared.
}
\description{
Returns the genome assembly declared for the regions of a RegionSetDE object.
}
\examples{
counts <- loadExampleData("counts", verbose = FALSE)
genomeAssembly(counts)

}
\seealso{
\code{\link{loadRegions}}
}
\author{
Sebastian Gregoricchio
}
