% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{regionRanges}
\alias{regionRanges}
\alias{regionRanges,RegionSetDE-method}
\title{regionRanges}
\usage{
regionRanges(object)

\S4method{regionRanges}{RegionSetDE}(object)
}
\arguments{
\item{object}{\code{RegionSetDE} object returned by \code{\link{loadRegions}}, \code{\link{splitLoadRegions}} or \code{\link{loadConsensusPeaks}}.}
}
\value{
A named \code{GRangesList} with the regions of every set, together with their metadata columns.
}
\description{
Returns the region sets of a \code{RegionSetDE} object as a named \code{GRangesList}, one element per set.
}
\examples{
regionTable <- loadExampleData("regions", verbose = FALSE)

regions <- splitLoadRegions(GenomicRanges::makeGRangesFromDataFrame(regionTable, keep.extra.columns = TRUE),
                            splitBy = "setName", genomeAssembly = "rn4", verbose = FALSE)

lengths(regionRanges(regions))
head(regionRanges(regions)$promoterCpG, 3)

}
\seealso{
\code{\link{regionSetNames}}, \code{\link{resultRanges}}
}
\author{
Sebastian Gregoricchio
}
