% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDE-package.R
\name{RegionSetDE-exampleData}
\alias{RegionSetDE-exampleData}
\title{RegionSetDE example data}
\description{
Description of the example objects installed under \code{inst/extdata}
and of the dataset they come from. The objects are read with \code{\link{loadExampleData}},
which is where the individual formats are documented.
}
\seealso{
\code{\link{loadExampleData}}
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
