% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RegionSetDE.counts-class}
\alias{RegionSetDE.counts-class}
\title{RegionSetDE.counts class}
\description{
S4 class storing the read counts computed over a collection of region sets. It extends \code{RangedSummarizedExperiment}, therefore \code{assay}, \code{colData}, \code{rowRanges} and the subsetting operators behave as usual, while the filters applied upstream remain accessible in the inherited provenance slots.
}
\section{Slots}{

\describe{
\item{\code{counting.level}}{String indicating whether the rows are regions (\code{"region"}) or tiles of a region (\code{"tile"}). It is the same vocabulary the \code{RegionSetDE.fit} and \code{RegionSetDE.results} classes use, since the value travels from here into both of them.}
}}

\author{
Sebastian Gregoricchio
}
