% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RegionSetDE.resultsList-class}
\alias{RegionSetDE.resultsList-class}
\title{RegionSetDE.resultsList class}
\description{
S4 class holding the outcome of several contrasts run on the same fit. Since all of them come from one model, the design, the offsets and the dispersion are shared and the contrasts can be compared with each other directly. Every function that takes a \code{RegionSetDE.results} object also takes this one, together with a \code{contrast} argument naming which of them to use.
}
\section{Slots}{

\describe{
\item{\code{results}}{Named list of \code{RegionSetDE.results} objects.}

\item{\code{contrasts}}{Character vector with the names of the contrasts, in the order they were run.}
}}

\author{
Sebastian Gregoricchio
}
