% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RegionSetDE.setResultsList-class}
\alias{RegionSetDE.setResultsList-class}
\title{RegionSetDE.setResultsList class}
\description{
S4 class holding the set level outcome of several contrasts run on the same fit. Every function that takes a \code{RegionSetDE.setResults} object also takes this one, together with a \code{contrast} argument naming which of them to use.
}
\section{Slots}{

\describe{
\item{\code{results}}{Named list of \code{RegionSetDE.setResults} objects.}

\item{\code{contrasts}}{Character vector with the names of the contrasts, in the order they were run.}
}}

\author{
Sebastian Gregoricchio
}
