% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMethods.R
\name{resultRanges}
\alias{resultRanges}
\alias{resultRanges,RegionSetDE.results-method}
\alias{resultRanges,RegionSetDE.resultsList-method}
\title{resultRanges}
\usage{
resultRanges(results)

\S4method{resultRanges}{RegionSetDE.results}(results)

\S4method{resultRanges}{RegionSetDE.resultsList}(results)
}
\arguments{
\item{results}{\code{RegionSetDE.results} object.}
}
\value{
A \code{GRanges} with one element per region.
}
\description{
Returns the coordinates of a \code{RegionSetDE.results} object with the statistics attached as metadata columns, ready to be exported as a BED-like file.
}
\examples{
fit <- loadExampleData("fit", verbose = FALSE)
results <- testRegions(fit, contrast = c("condition", "SHR", "BN"), verbose = FALSE)

resultRanges(results)

}
\author{
Sebastian Gregoricchio
}
