% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE-method}
\alias{show,RegionSetDE-method}
\title{show method for RegionSetDE objects}
\usage{
\S4method{show}{RegionSetDE}(object)
}
\arguments{
\item{object}{A \code{RegionSetDE} object.}
}
\value{
Prints a summary to the console.
}
\description{
Prints a summary of the region sets stored in a \code{RegionSetDE} object.
}
\author{
Sebastian Gregoricchio
}
