% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.fit-method}
\alias{show,RegionSetDE.fit-method}
\title{show method for RegionSetDE.fit}
\usage{
\S4method{show}{RegionSetDE.fit}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.fit} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints a summary of a fitted model.
}
\author{
Sebastian Gregoricchio
}
