% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.results-method}
\alias{show,RegionSetDE.results-method}
\title{show method for RegionSetDE.results}
\usage{
\S4method{show}{RegionSetDE.results}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.results} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints a summary of a per-region contrast.
}
\author{
Sebastian Gregoricchio
}
