% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.resultsList-method}
\alias{show,RegionSetDE.resultsList-method}
\title{show method for RegionSetDE.resultsList}
\usage{
\S4method{show}{RegionSetDE.resultsList}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.resultsList} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints the contrasts held by the object and how many regions change in each of them.
}
\author{
Sebastian Gregoricchio
}
