% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.setResults-method}
\alias{show,RegionSetDE.setResults-method}
\title{show method for RegionSetDE.setResults}
\usage{
\S4method{show}{RegionSetDE.setResults}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.setResults} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints a summary of a set level test.
}
\author{
Sebastian Gregoricchio
}
