% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.setResultsList-method}
\alias{show,RegionSetDE.setResultsList-method}
\title{show method for RegionSetDE.setResultsList}
\usage{
\S4method{show}{RegionSetDE.setResultsList}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.setResultsList} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints the contrasts held by the object and the sets tested in each of them.
}
\author{
Sebastian Gregoricchio
}
