% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.setScores-method}
\alias{show,RegionSetDE.setScores-method}
\title{show method for RegionSetDE.setScores}
\usage{
\S4method{show}{RegionSetDE.setScores}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.setScores} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints a summary of the region set scores and of the comparisons between them.
}
\examples{
counts <- loadExampleData("counts", verbose = FALSE)
scoreRegionSets(counts, verbose = FALSE)

}
\author{
Sebastian Gregoricchio
}
