% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,RegionSetDE.universe-method}
\alias{show,RegionSetDE.universe-method}
\title{show method for RegionSetDE.universe}
\usage{
\S4method{show}{RegionSetDE.universe}(object)
}
\arguments{
\item{object}{\code{RegionSetDE.universe} object.}
}
\value{
Prints the summary to the console.
}
\description{
Prints a summary of a comparison universe.
}
\author{
Sebastian Gregoricchio
}
