% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{updateObject-RegionSetDE}
\alias{updateObject-RegionSetDE}
\alias{updateObject,RegionSetDE.counts-method}
\alias{updateObject,RegionSetDE.fit-method}
\alias{updateObject,RegionSetDE.results-method}
\alias{updateObject,RegionSetDE.setResults-method}
\alias{updateObject,RegionSetDE.setScores-method}
\alias{updateObject,RegionSetDE.resultsList-method}
\alias{updateObject,RegionSetDE.setResultsList-method}
\title{updateObject methods for the RegionSetDE classes}
\usage{
\S4method{updateObject}{RegionSetDE}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.counts}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.fit}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.results}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.setResults}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.setScores}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.resultsList}(object, ..., verbose = FALSE)

\S4method{updateObject}{RegionSetDE.setResultsList}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{An object of one of the RegionSetDE classes.}

\item{...}{Passed on to the method of \code{SummarizedExperiment} for a counts object, ignored otherwise.}

\item{verbose}{Logical value to indicate whether the messages must be printed. Default: \code{FALSE}.}
}
\value{
The object, updated to the current definition of its class. An object already up to date comes back as it was.
}
\description{
Bring an object saved by an earlier version of RegionSetDE to the current definition of its class, as \code{BiocGenerics::updateObject} does for the Bioconductor classes. An object saved before the \code{greylist} slot existed gets it, filled with the greylist of the consensus when the regions were built by \code{\link{loadConsensusPeaks}}, and empty otherwise. A fit or a result updates the counts it carries as well, and a list of results every result in it.
}
\details{
A greylist applied with \code{\link{applyGreylist}} before the slot existed left only its size in the object, so the update cannot recover it: apply it again, or pass it to the \code{greylist} argument of \code{\link{countReads}}.
}
\examples{
counts <- loadExampleData("counts", verbose = FALSE)
counts <- BiocGenerics::updateObject(counts, verbose = TRUE)

}
\author{
Sebastian Gregoricchio
}
