% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsUtils.R
\name{.makeParallelParam}
\alias{.makeParallelParam}
\title{.makeParallelParam}
\usage{
.makeParallelParam(nThreads = 1)
}
\arguments{
\item{nThreads}{Number of threads. Default: \code{1}.}
}
\value{
A \code{BiocParallelParam} object.
}
\description{
Builds the \code{BiocParallel} back end matching the number of requested threads and the operating system.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
