% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRegions.R
\name{.testDESeq2}
\alias{.testDESeq2}
\title{.testDESeq2}
\usage{
.testDESeq2(fit, contrastVector, lfcThreshold = 0)
}
\arguments{
\item{fit}{\code{RegionSetDE.fit} object.}

\item{contrastVector}{Numeric vector with the contrast.}

\item{lfcThreshold}{Numeric value with the log2 fold change of the null hypothesis.}
}
\value{
A data.frame with the \code{log2FC}, \code{average.signal}, \code{stat} and \code{p.value} columns.
}
\description{
Runs the Wald test on a \code{DESeq2} fit.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
