% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRegions.R
\name{.testVoom}
\alias{.testVoom}
\title{.testVoom}
\usage{
.testVoom(fit, contrastVector, lfcThreshold = 0)
}
\arguments{
\item{fit}{\code{RegionSetDE.fit} object.}

\item{contrastVector}{Numeric vector with the contrast.}

\item{lfcThreshold}{Numeric value with the log2 fold change of the null hypothesis.}
}
\value{
A data.frame with the \code{log2FC}, \code{average.signal}, \code{stat} and \code{p.value} columns.
}
\description{
Runs the moderated t test on a \code{limma} fit.
}
\author{
Sebastian Gregoricchio
}
\keyword{internal}
