% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMethods.R
\name{fitCounts}
\alias{fitCounts}
\alias{fitCounts,RegionSetDE.fit-method}
\title{fitCounts}
\usage{
fitCounts(fit)

\S4method{fitCounts}{RegionSetDE.fit}(fit)
}
\arguments{
\item{fit}{\code{RegionSetDE.fit} object.}
}
\value{
A \code{RegionSetDE.counts} object.
}
\description{
Returns the \code{RegionSetDE.counts} object on which a model was fitted.
}
\examples{
fit <- loadExampleData("fit", verbose = FALSE)

# The counts the model was fitted on, normalisation and filtering included
fitCounts(fit)

# The underlying engine object, here an edgeR fit
class(fitObject(fit))


}
\author{
Sebastian Gregoricchio
}
