% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMethods.R
\name{fitObject}
\alias{fitObject}
\alias{fitObject,RegionSetDE.fit-method}
\title{fitObject}
\usage{
fitObject(fit)

\S4method{fitObject}{RegionSetDE.fit}(fit)
}
\arguments{
\item{fit}{\code{RegionSetDE.fit} object.}
}
\value{
The engine specific fit object.
}
\description{
Returns the object produced by the engine, untouched, so that any function of \code{edgeR}, \code{limma}, \code{variancePartition} or \code{DESeq2} can be applied to it.
}
\author{
Sebastian Gregoricchio
}
