% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExampleData.R
\name{loadExampleData}
\alias{loadExampleData}
\title{Load the packaged NF-YA example objects}
\usage{
loadExampleData(
  dataset = c("readCounts", "backgroundCounts", "regions", "regionSets", "blacklist",
    "sampleSheet", "buildMetadata")
)
}
\arguments{
\item{dataset}{Character, the object to load. One of \code{"readCounts"},
\code{"backgroundCounts"}, \code{"regions"}, \code{"regionSets"},
\code{"blacklist"}, \code{"sampleSheet"} or \code{"buildMetadata"}.}
}
\value{
The requested object.
}
\description{
Loads one of the precomputed objects shipped in \code{inst/extdata}. These
were built from the NF-YA ChIP-seq libraries distributed by
\pkg{chipseqDBData}, restricted to chr17 and chr19 of mm10. See
\code{inst/scripts/make-data.R} for how each object was generated.
}
\examples{
regionSets <- loadExampleData("regionSets")
head(regionSets)

}
\author{
Sebastian Gregoricchio
}
