% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RegionSetDE-class}
\alias{RegionSetDE-class}
\title{RegionSetDE class}
\description{
S4 class collecting a group of genomic region sets together with the filters applied to them. The regions are stored as a \code{GRangesList}, so that any Bioconductor operation remains available through the \code{regions} slot.
}
\section{Slots}{

\describe{
\item{\code{regions}}{\code{GRangesList} containing the region sets, one element per set.}
}}

\author{
Sebastian Gregoricchio
}
