% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMethods.R
\name{regionSetNames}
\alias{regionSetNames}
\alias{regionSetNames,RegionSetDE.results-method}
\alias{regionSetNames,RegionSetDE.fit-method}
\title{regionSetNames}
\usage{
regionSetNames(object)

\S4method{regionSetNames}{RegionSetDE.results}(object)

\S4method{regionSetNames}{RegionSetDE.fit}(object)
}
\arguments{
\item{object}{\code{RegionSetDE}, \code{RegionSetDE.counts}, \code{RegionSetDE.fit} or \code{RegionSetDE.results} object.}
}
\value{
A character vector.
}
\description{
Returns the names of the region sets held by an object of the package.
}
\author{
Sebastian Gregoricchio
}
