% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{[[,RegionSetDE.setResultsList,ANY,ANY-method}
\alias{[[,RegionSetDE.setResultsList,ANY,ANY-method}
\alias{$,RegionSetDE.setResultsList-method}
\alias{names,RegionSetDE.setResultsList-method}
\alias{length,RegionSetDE.setResultsList-method}
\alias{[[,RegionSetDE.resultsList,ANY,ANY-method}
\alias{$,RegionSetDE.resultsList-method}
\alias{names,RegionSetDE.resultsList-method}
\alias{length,RegionSetDE.resultsList-method}
\title{Accessors of RegionSetDE.resultsList}
\usage{
\S4method{[[}{RegionSetDE.setResultsList,ANY,ANY}(x, i, j, ...)

\S4method{$}{RegionSetDE.setResultsList}(x, name)

\S4method{names}{RegionSetDE.setResultsList}(x)

\S4method{length}{RegionSetDE.setResultsList}(x)

\S4method{[[}{RegionSetDE.resultsList,ANY,ANY}(x, i, j, ...)

\S4method{$}{RegionSetDE.resultsList}(x, name)

\S4method{names}{RegionSetDE.resultsList}(x)

\S4method{length}{RegionSetDE.resultsList}(x)
}
\arguments{
\item{x}{\code{RegionSetDE.resultsList} object.}

\item{i}{String with the name of a contrast, or its position.}

\item{...}{Not used.}

\item{name}{String with the name of a contrast.}
}
\value{
A \code{RegionSetDE.results} object, or a character vector for \code{names}.
}
\description{
Extract one contrast, or the names of the contrasts, from a \code{RegionSetDE.resultsList} object.
}
\author{
Sebastian Gregoricchio
}
