% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsMethods.R
\name{resultsTable}
\alias{resultsTable}
\alias{resultsTable,RegionSetDE.results-method}
\title{resultsTable}
\usage{
resultsTable(results)

\S4method{resultsTable}{RegionSetDE.results}(results)
}
\arguments{
\item{results}{\code{RegionSetDE.results} object.}
}
\value{
A data.frame with one row per region.
}
\description{
Returns the per-region table of a \code{RegionSetDE.results} object.
}
\author{
Sebastian Gregoricchio
}
