% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSamples.R
\name{splitSamples}
\alias{splitSamples}
\title{splitSamples}
\usage{
splitSamples(
  counts,
  by,
  dropNormalization = TRUE,
  minSamples = 1,
  verbose = TRUE
)
}
\arguments{
\item{counts}{\code{RegionSetDE.counts} object.}

\item{by}{String, or character vector, with the names of the \code{colData} columns defining the groups. When more than one is given the levels are combined.}

\item{dropNormalization}{Logical value to indicate whether the normalisation stored in the object must be discarded in each piece. Default: \code{TRUE}.}

\item{minSamples}{Numeric value with the minimum number of samples a group must contain to be returned. Default: \code{1}.}

\item{verbose}{Logical value to indicate whether the messages must be printed. Default: \code{TRUE}.}
}
\value{
A named list of \code{RegionSetDE.counts} objects.
}
\description{
Splits a \code{RegionSetDE.counts} object into a list of objects, one per level of a column of the \code{colData}. Convenient when several marks or assays have been counted together over the same regions and each of them needs its own normalisation and its own fit.
}
\examples{
counts <- loadExampleData("counts", verbose = FALSE)

byStrain <- splitSamples(counts, by = "condition", verbose = FALSE)
names(byStrain)
byStrain[[1]]

}
\seealso{
\code{\link{selectSamples}}, \code{\link{fitRegions}}
}
\author{
Sebastian Gregoricchio
}
